\name{combTraits}
\alias{combTraits}
\title{Combines different traits into one}
\usage{
  combTraits(data, unified, targets)
}
\arguments{
  \item{data}{Central database}

  \item{targets}{Vector of columns names to be combined}

  \item{unified}{Name of final unified trait}
}
\value{
  Central database with additional unified trait column
}
\description{
  Combines different traits into one
}

