\name{createNullTran}
\alias{createNullTran}
\title{Creates objects containing WNODF and NODF for a list of matrices}
\usage{
  createNullTran(matrices, names, ...)
}
\arguments{
  \item{matrices}{A list of matrices}

  \item{names}{Tags to be pasted into each matrix object
  name}

  \item{...}{optional arguments to \code{nullMatsVegan}}
}
\value{
  assigns lists of null matrices to global environment
}
\description{
  Creates objects containing WNODF and NODF for a list of
  matrices
}

