\name{matNODF}
\alias{matNODF}
\title{Calculates NODF of a matrix}
\usage{
  matNODF(x, summary = FALSE, index = 3, ...)
}
\arguments{
  \item{x}{a matrix}

  \item{summary}{obtains summary for whole matrix, row or
  column only}

  \item{index}{if summary=TRUE, \code{index} is applied on
  the statistic list element of a obkect of class
  \code{nestednodf}. Default is 3, i.e., whole matrix NODF}

  \item{...}{optional arguments to \code{nestednodf}}
}
\value{
  NODF
}
\description{
  Calculates NODF of a matrix
}
\references{
  ????
}

