\name{myImagePlot}
\alias{myImagePlot}
\title{Plot binary and quantitative matrices with a color code (in {phaget4})}
\usage{
  myImagePlot(x, ...)
}
\arguments{
  \item{x}{matrix}

  \item{...}{whatever parameters a plot function may
  receive}
}
\value{
  colored matrix plot
}
\description{
  Plot binary and quantitative matrices with a color code
  (in {phaget4})
}
\details{
  function could be retrieved by
  'source("http://www.phaget4.org/R/myImagePlot.R")'
}
\references{
  citation(package = "phaget4")
}

