\name{nestObjects}
\alias{nestObjects}
\title{Creates objects containing WNODF and NODF for a list of matrices}
\usage{
  nestObjects(x, fgLevels, names, ...)
}
\arguments{
  \item{x}{A list of matrices}

  \item{fgLevels}{levels of traits combinations}

  \item{names}{Object tags to be pasted to each
  \code{fgLevels} which will results in the object names}

  \item{...}{optional arguments to \code{matNODF}}
}
\value{
  assigns objects to global environment
}
\description{
  Creates objects containing WNODF and NODF for a list of
  matrices
}

