\name{nullMatsMgen}
\alias{nullMatsMgen}
\title{Iterates over Mgen algorithm}
\usage{
  nullMatsMgen(m, iter, ...)
}
\arguments{
  \item{m}{A quantitative matrix}

  \item{iter}{Number of iteractions, i.e. number of random
  matrices to be created}

  \item{...}{Further arguments from \code{Mgen}}
}
\value{
  a list with iter random matrices
}
\description{
  Iterates over \code{Mgen} for quantitative matrices.
}

