\name{plotFuncDiff}
\alias{plotFuncDiff}
\title{Plots differences between random and original matrix with respect to a function}
\usage{
  plotFuncDiff(original, random, x, lim, ...)
}
\arguments{
  \item{x}{function used for comparison}

  \item{original}{Original matrix}

  \item{random}{One of the random matrices generated via
  \code{nullMatsVegan}}

  \item{...}{additional arguments to \code{plot}}
}
\value{
  A plot with the 1-1 line
}
\description{
  Plots differences between random and original matrix with
  respect to a function
}

