\name{summaryNODF}
\alias{summaryNODF}
\title{Calculates summary of NODFs across all random matrices}
\usage{
  summaryNODF(matList, ...)
}
\arguments{
  \item{matList}{list containing random matrices}

  \item{...}{optional arguments to \code{matNODF}}
}
\value{
  Summary statistics of NODF values across all random
  matrices in matList
}
\description{
  Calculates summary of NODFs across all random matrices
}
\references{
  ????
}

